encoding with fewer tests
Minor encoding speedup

+++ bmap.cpp
@@ -114,7 +114,7 @@ size_t BMap::unpack(Bitstream& s) {
                     q = code ? 0xff00 : 0x00ff;
                 else {
                     if (5 < code) // Need one more code bit
-                        code = static_cast<uint8_t>(s.get() | (code << 1));
+                        code = static_cast<uint8_t>(s.get() | (static_cast<uint64_t>(code) << 1));
                     s.pull(q, 7); // Need 7 bits
                     switch (code) { // Combo with one mixed byte
                     case 0b010:                         break; // 0b0010
